backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
psycopg2-binary
python-dotenv
requests
statsmodels
yfinance
